fixing mpi function linking
